PlayerPROCore: Don't open a bundle's resource fork on OS X: Depend on the plug-in to do it for itself, as many of the current plug-ins don't even have a resource fork. Change the second argument for CallImportPlug to an int. Update the BeOS port.

+++ MADDriver.source/Lin-PlugImport.c
@@ -47,11 +47,11 @@ MADErr PPMADInfoFile(const char *AlienFile, MADInfoRec *InfoRec)
 }
 
 MADErr CallImportPlug(MADLibrary	*inMADDriver,
-					  short			PlugNo,			// CODE du plug
+					  int			PlugNo,			// CODE du plug
 					  MADFourChar	order,
 					  char			*AlienFile,
 					  MADMusic		*theNewMAD,
-					  MADInfoRec		*info)
+					  MADInfoRec	*info)
 {
 	MADDriverSettings driverSettings = {0};
